Add multiple test files
